Add support for refrigerator
Add entities for refrigerator, such as temperatures and defrosting
status. Some settings should be configurable in the Panasonic App,
but are currently read-only in this commit due the complexity. These
are only tested with a Panasonic NR-D611XGS refrigerator, and might
not work with other models.

Also add a CO2 footprint sensor that might be available for other
appliances, which is similar to the energy consumption sensor.